Project UI/requirements.txt
flask
numpy
pillow
h5py
gunicorn
